requirements.txt
Django==5.2.4
djangorestframework==3.16.0
psycopg2==2.9.10
beautifulsoup4==4.13.4
asgiref~=3.9.1
requests~=2.32.4
python-decouple~=3.8